create extension serve namespace
