+++ lib/project_types/extension/models/specification_handlers/theme_app_extension.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 require "base64"
 require "json"
-require "shopify_cli/theme/app_extension_dev_server"
+require "shopify_cli/theme/extension/dev_server"
 
 module Extension
   module Models
@@ -78,7 +78,7 @@ module Extension
           # port = options[:port]
           root = options[:context]&.root
 
-          ShopifyCLI::Theme::DevServer::AppExtensionDevServer.start(@ctx, root)
+          ShopifyCLI::Theme::Extension::DevServer.start(@ctx, root)
         end
 
         private
