+++ lib/shopify_cli/theme/dev_server/errors.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+module ShopifyCLI
+  module Theme
+    module DevServer
+      # Errors
+      class Error < StandardError; end
+      class AddressBindingError < StandardError; end
+    end
+  end
+end
