+++ lib/shopify_cli/theme/dev_server/proxy.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 require_relative "base_proxy"
-require_relative "proxy/theme_template_param_builder"
+require_relative "proxy_helpers/template_param_builder"
 
 module ShopifyCLI
   module Theme
     module DevServer
       # Created for separation of unit tests
-      class ThemeProxy < BaseProxy
+      class Proxy < BaseProxy
         def initialize(ctx, theme:, syncer:)
           super(ctx)
           @theme = theme
@@ -19,7 +19,7 @@ module ShopifyCLI
         private
 
         def build_replacement_param(env)
-          ThemeTemplateParamBuilder.new
+          ProxyHelpers::TemplateParamBuilder.new
             .with_core_endpoints(@core_endpoints)
             .with_syncer(@syncer)
             .with_theme(@theme)
