+++ test/shopify-cli/theme/dev_server/proxy_helpers/template_param_builder_test.rb
@@ -1,16 +1,16 @@
 # frozen_string_literal: true
 
 require "test_helper"
-require "shopify_cli/theme/dev_server/proxy/theme_template_param_builder"
+require "shopify_cli/theme/dev_server/proxy_helpers/template_param_builder"
 
 module ShopifyCLI
   module Theme
     module DevServer
-      class BaseProxy
-        class ThemeTemplateParamBuilderTest < Minitest::Test
+      module ProxyHelpers
+        class TemplateParamBuilderTest < Minitest::Test
           def setup
             super
-            @param_builder = ThemeTemplateParamBuilder.new
+            @param_builder = TemplateParamBuilder.new
           end
 
           def test_empty_build
