+++ test/shopify-cli/theme/dev_server/watcher_test.rb
@@ -1,19 +1,18 @@
 # frozen_string_literal: true
 require "test_helper"
-require "shopify_cli/theme/theme_dev_server"
-require "shopify_cli/theme/dev_server/theme_watcher"
+require "shopify_cli/theme/dev_server"
 
 module ShopifyCLI
   module Theme
     module DevServer
-      class ThemeWatcherTest < Minitest::Test
+      class WatcherTest < Minitest::Test
         def setup
           super
           @root = ShopifyCLI::ROOT + "/test/fixtures/theme"
           @ctx = TestHelpers::FakeContext.new(root: @root)
           @theme = Theme.new(@ctx, root: @root)
           @syncer = stub("Syncer", enqueue_uploads: true, enqueue_updates: true)
-          @watcher = ThemeWatcher.new(@ctx, theme: @theme, syncer: @syncer)
+          @watcher = Watcher.new(@ctx, theme: @theme, syncer: @syncer)
         end
 
         def test_upload_files_when_changed
